refactor: split graphql code into app and server files
