+++ package.json
@@ -16,6 +16,8 @@
   },
   "scripts": {
     "server": "node ./src/server.js",
-    "dev": "nodemon ./src/server.js"
+    "dev": "nodemon ./src/server.js",
+    "test": "jest --coverage",
+    "test:watch": "jest --watch --coverage"
   }
 }
