+++ src/__tests__/appTest.js
@@ -1,0 +1,5 @@
+describe("Example test", () => {
+    test("It should pass", async () => {
+        expect(1).toBe(1);
+    })
+})
